requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
SentenceTransformers==2.2.2
scikit-learn==1.3.1
numpy==1.24.3
